Refactor data quality and processing modules, consolidate aliases
