+++ src/feature_engineering/cross_timeframe_analysis_pipeline.py
@@ -66,7 +66,7 @@ class MathValidation:
         return safe_power(x, y, default)
 
 # Import data quality utilities from data_quality
-from src.utils.data.quality.enhanced_data_quality_validator import QualityResult, EnhancedDataQualityValidator
+from src.utils.data.quality.data_quality import QualityResult, DataQualityFramework as EnhancedDataQualityValidator
 
 # Simple placeholder classes for missing functionality
 class DataQualityUtilities:
