+++ src/utils/data/__init__.py
@@ -55,8 +55,61 @@ from .unified_data_utils import (
     unified_data_utils
 )
 
-# Import backwards compatibility aliases
-from .backwards_compatibility import *
+# Backwards compatibility aliases - integrated directly
+class DataFrameValidator:
+    """Backwards compatibility alias for DataFrameValidator."""
+    def __init__(self):
+        self.processor = DataProcessor()
+    
+    def validate_dataframe(self, df, **kwargs):
+        return self.processor.validate_and_fix_data_quality(df, **kwargs)
+
+class DataFrameCleaner:
+    """Backwards compatibility alias for DataFrameCleaner."""
+    def __init__(self):
+        self.processor = DataProcessor()
+        self.cleaner = DataCleaner()
+    
+    def clean_dataframe(self, df, **kwargs):
+        return self.processor.validate_and_fix_data_quality(df, **kwargs)
+
+class DataFrameTransformer:
+    """Backwards compatibility alias for DataFrameTransformer."""
+    def __init__(self):
+        self.processor = DataProcessor()
+    
+    def transform_dataframe(self, df, **kwargs):
+        return self.processor.regularize_timestamps(df, **kwargs)
+
+# Backwards compatibility functions
+def validate_dataframe(df, **kwargs):
+    """Backwards compatibility function for validate_dataframe."""
+    processor = DataProcessor()
+    return processor.validate_and_fix_data_quality(df, **kwargs)
+
+def clean_dataframe(df, **kwargs):
+    """Backwards compatibility function for clean_dataframe."""
+    processor = DataProcessor()
+    return processor.validate_and_fix_data_quality(df, **kwargs)
+
+def transform_dataframe(df, **kwargs):
+    """Backwards compatibility function for transform_dataframe."""
+    processor = DataProcessor()
+    return processor.regularize_timestamps(df, **kwargs)
+
+# Backwards compatibility for old class names
+DataFormattingFramework = DataQualityFramework
+DataFormat = QualityResult
+ColumnNamingConvention = DataSchema
+
+# Backwards compatibility for old function names
+data_preprocessing = preprocess_data_for_multi_timeframe
+data_processing_utils = optimize_dataframe_dtypes
+enhanced_data_operations = apply_feature_specific_optimization
+
+# Backwards compatibility for old class instances
+DataLoader = DataStreamingManager
+OptimizedDataManager = DataProcessor
 
 __all__ = [
     # Data Quality Framework
@@ -104,5 +157,23 @@ __all__ = [
     
     # Unified Interface
     'UnifiedDataUtils',
-    'unified_data_utils'
+    'unified_data_utils',
+    
+    # Backwards Compatibility Classes
+    'DataFrameValidator',
+    'DataFrameCleaner',
+    'DataFrameTransformer',
+    'DataFormattingFramework',
+    'DataFormat',
+    'ColumnNamingConvention',
+    'DataLoader',
+    'OptimizedDataManager',
+    
+    # Backwards Compatibility Functions
+    'validate_dataframe',
+    'clean_dataframe',
+    'transform_dataframe',
+    'data_preprocessing',
+    'data_processing_utils',
+    'enhanced_data_operations'
 ]
