+++ src/utils/data/quality/advanced_quality_metrics.py
@@ -9,8 +9,6 @@ from dataclasses import dataclass, field
 from scipy import stats
 from ..logger import system_logger
 from src.utils.pipeline_standards import PipelineStandards
-
-from ..logger import system_logger
 import logging
 import time
